Added optimization of re-renders using memo, added tanstack query devtools, bug fixes

+++ backend/app/repository/user_repository.py
@@ -61,7 +61,10 @@ class UserRepository:
     
   async def delete_user(self, user_id: int):
     user = await self.get_user_by_id(user_id)
-    print('user', user)
+
+    if user is None:
+      raise AppError(404, f"Користувач за ID {user_id} не знайдено")
+    print(user)
     try:
       await self.db.delete(user)
       await self.db.commit()
